+++ deps.ts
@@ -1,1 +1,1 @@
-export { join } from "https://deno.land/std@0.99.0/path/mod.ts";
+// :D