+++ examples/restrictedRead/main.ts
@@ -1,24 +1,36 @@
-import { importw } from "../../mod.ts";
+import { importw, release } from "../../mod.ts";
 
-const exampleModPath = new URL("./exampleMod.ts", import.meta.url).pathname;
+export async function main() {
+  const exampleModPath = new URL("./exampleMod.ts", import.meta.url).pathname;
 
-// Import `denoCwd` from a Worker with Deno namespace enabled, but read permission disallowed
-const { denoCwd } = await importw(
-  "./examples/restrictedRead/exampleMod.ts",
-  {
-    name: "exampleWorkerWithDenoNamespaceButNoRead",
-    deno: {
-      namespace: true,
-      permissions: {
-        read: [exampleModPath],
+  // Import `denoCwd` from a Worker with Deno namespace enabled, but read permission disallowed
+  const { denoCwd, [release]: terminate } = await importw(
+    "./examples/restrictedRead/exampleMod.ts",
+    {
+      name: "permissionRestrictedWorker",
+      deno: {
+        namespace: true,
+        permissions: {
+          read: [exampleModPath],
+        },
       },
     },
-  },
-);
+  );
 
-/**
- * Expect to get a read permission error:
- *
- * `error: Uncaught (in promise) PermissionDenied: read access to <CWD>, run again with the --allow-read flag`
- */
-await denoCwd(); // BOOM!
+  /**
+   * Expect to get a read permission error:
+   * `PermissionDenied: read access to <CWD>, run again with the --allow-read flag`
+   */
+  try {
+    await denoCwd(); // BOOM!
+  } catch (e) {
+    console.log("BOOM! 💥 ");
+    console.error(e);
+  }
+
+  await terminate();
+}
+
+if (import.meta.main) {
+  await main();
+}