+++ mod.ts
@@ -1,2 +1,4 @@
-export { importw, workerSymbol } from "./src/importw.ts";
-export { importw as default } from "./src/importw.ts";
+export { importw, worker } from "./src/importw.ts";
+
+// @deno-types="./src/comlink.d.ts"
+export { releaseProxy as release } from "./src/comlink.js";