+++ src/notImplemented.ts
@@ -1,5 +1,0 @@
-export function notImplemented(): void {
-  console.warn(
-    "This feature has not been fully implemented, usage may result in undesired behaviour",
-  );
-}