+++ src/protocol.d.ts
@@ -1,10 +1,5 @@
-// deno-lint-ignore-file no-explicit-any
+// deno-lint-ignore-file
 /**
- * This is a limited port of comlink
- * (https://github.com/GoogleChromeLabs/comlink) for Deno.
- *
- * comlink originally licensed as follows:
- *
  * Copyright 2019 Google Inc. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,127 +11,89 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import { MessagePortPolyfill } from "./messagePortPolyfill.js";
-
-// TODO: use built-in type once available
-// REF: https://developer.mozilla.org/en-US/docs/Web/API/Transferable
-export type Transferable = ArrayBuffer | MessagePortPolyfill;
-
 export interface EventSource {
   addEventListener(
     type: string,
     listener: EventListenerOrEventListenerObject,
-    options?: unknown,
+    options?: {},
   ): void;
   removeEventListener(
     type: string,
     listener: EventListenerOrEventListenerObject,
-    options?: unknown,
+    options?: {},
   ): void;
 }
-
 export interface PostMessageWithOrigin {
   postMessage(
     message: any,
     targetOrigin: string,
-    transferables?: Transferable[],
+    transfer?: Transferable[],
   ): void;
 }
-
 export interface Endpoint extends EventSource {
-  postMessage(message: any, transferables?: Transferable[]): void;
+  postMessage(message: any, transfer?: Transferable[]): void;
   start?: () => void;
 }
-
-export const WireValueMap = {
-  RAW: 0,
-  PROXY: 1,
-  THROW: 2,
-  HANDLER: 3,
-};
-
-export const enum WireValueType {
-  RAW,
-  PROXY,
-  THROW,
-  HANDLER,
+export declare const enum WireValueType {
+  RAW = "RAW",
+  PROXY = "PROXY",
+  THROW = "THROW",
+  HANDLER = "HANDLER",
 }
-
 export interface RawWireValue {
   id?: string;
   type: WireValueType.RAW;
-  value: unknown;
+  value: {};
 }
-
 export interface HandlerWireValue {
   id?: string;
   type: WireValueType.HANDLER;
   name: string;
   value: unknown;
 }
-
-export type WireValue = RawWireValue | HandlerWireValue;
-
-export type MessageID = string;
-
-export const MessageMap = {
-  GET: 0,
-  SET: 1,
-  APPLY: 2,
-  CONSTRUCT: 3,
-  ENDPOINT: 4,
-  RELEASE: 5,
-};
-
-export const enum MessageType {
-  GET,
-  SET,
-  APPLY,
-  CONSTRUCT,
-  ENDPOINT,
-  RELEASE,
+export declare type WireValue = RawWireValue | HandlerWireValue;
+export declare type MessageID = string;
+export declare const enum MessageType {
+  GET = "GET",
+  SET = "SET",
+  APPLY = "APPLY",
+  CONSTRUCT = "CONSTRUCT",
+  ENDPOINT = "ENDPOINT",
+  RELEASE = "RELEASE",
 }
-
 export interface GetMessage {
   id?: MessageID;
   type: MessageType.GET;
   path: string[];
 }
-
 export interface SetMessage {
   id?: MessageID;
   type: MessageType.SET;
   path: string[];
   value: WireValue;
 }
-
 export interface ApplyMessage {
   id?: MessageID;
   type: MessageType.APPLY;
   path: string[];
   argumentList: WireValue[];
 }
-
 export interface ConstructMessage {
   id?: MessageID;
   type: MessageType.CONSTRUCT;
   path: string[];
   argumentList: WireValue[];
 }
-
 export interface EndpointMessage {
   id?: MessageID;
   type: MessageType.ENDPOINT;
 }
-
 export interface ReleaseMessage {
   id?: MessageID;
   type: MessageType.RELEASE;
   path: string[];
 }
-
-export type Message =
+export declare type Message =
   | GetMessage
   | SetMessage
   | ApplyMessage